Completely overhauled comparison system. - 'doComparison' method defined by ISupportsComparison interface is   now removed - along with the interface. - Equality: All value objects support basic == and != comparison.   - The base general implementation is located in Value::isEqualTo($r)     method.   - Any value class can extend this method's logic with its own     implementation. - Relation: Values can optionally support additional >, <, >=, <= comparison.   - The base general implementation is located in Value::hasRelationTo($op, $r)     method.   - Any value class can extend this method's logic with its own     implementation-
Basic equality:

When comparing A == B or A != B. Primi will invoke
A::isEqualTo(B). If this method returns true/false, the
comparison is deemed successfull. If this method returns null, it means
that the value didn't know how to resolve the comparison. In that case,
Primi will swap operands and try to do reverse comparison.

The opposite results of == vs != operators are resolved automatically
by Primi.

Basic relation:

When comparing A > B, A < B, A >= B or A <= B. Primi will invoke
A::hasRelationTo(operator, B). If this method returns true/false, the
comparison is deemed successfull. If this method returns null, error is thrown
and no operand swapping is performed.