Refactoring of shared logical/arithmetic/comparison handlers.
- Short-circuiting finally works correctly!
  - Only AND/OR operations are short circuited (as Python does it).
- (BC break) OR expression now returns the first non-falsey value it
  encounters (also as Python does it).
  - Previously the result was either simple true or false (as PHP does it),
    but this new behavior is much more useful.
    - For example this can now be used:
      - Expression 'subtract(2, 2) or 3' - now returns 3 instead of true.
- Chained comparisons still work fine (any maybe better).
  - Chained comparisons can be these: ==,!=,>,<,>=,<=.
    - For example: '3 > 2 > 1' returns true:
      - "3 is greater than 2 and also 2 is greater than 1".
    - Another example: '3 > 2 == 2' returns true:
      - "3 is greater than 2 and also 2 equals to 2".

Helper function yield_left_to_right which returns a generator is now
used to iterate over operands and operators when evaluating
logical/arithmetic/comparison AST nodes.

The generator yields just the first operand at first, then it continues
to yield tuples of [operator, operand] representing the rest of the
expression.

By yielding the first operand first (before any of the AST nodes
representing other operators/operands are executed - to not to evaluate
them too soon) we are able to implement proper short circuiting, which was
pretty much broken in previous versions of Primi.

- Interfaces allowing +,-,*,/ now defined nullable types.