Support for PHP 8.1
Removed ArrayAccess interface from MapContainer. Access to the container
is now done via standard functions instead of PHP magic ones.

This gets rid of deprecation errors for having untyped parameters
to, for example, ArrayAccess' method "offsetGet($key)".
We could get rid of by adding "mixed" typehint to the parameter, but that
is not supported by PHP 7.4, which we still want, so...

It's easier to get rid of the interface, as the MapContainer is used
only internally.

See:
- https://wiki.php.net/rfc/internal_method_return_types
- https://php.watch/versions/8.1/ReturnTypeWillChange

Also updated phpcb dev-dependency to 0.5.1 (also supports PHP 8.1).