Implemented "iteratorfactory" objects - range() and enumerate() return them.
This way we can have repeatable-iterable objects returned from those
functions.

For example, we want "range()" function (implemented in PHP) to return
not only a one-time iterator, but an iterable that allows multiple uses
as an iterator. Thus, PHP code handling the "range()" builtin Primi function
will return this "iteratorfactory" with a callable acting as a factory
function embedded within it.

Then, every time someone requests an iterator from this "iteratorfactory",
the factory callable will be executed and a brand-new generator is then
used as the actual internal iterator which then yields the expected values.

- GeneratorValue replaced with new IteratorFactoryValue.