Making sure Primi is able to act as a sandbox.
Original idea of this project was to allow sandboxing user-provided logic
for other PHP apps.

Since then Primi got more complex and powerful, and now it can be used
as a standalone scripting language. But we want to retain the
sandboxing abilities for situations where other projects (PHP apps) want
to use Primi as a sandbox for third-party scripts.

Sandbox is enabled by default in all runtimes if Primi is not running
as CLI program.

Sandboxing mode can also be specified via modified Config object.

Primi modules implemented in PHP (i.e. native modules) can use
the "DeniedInSandboxTrait" which will disallow importing that module
when in sandbox mode.