Finally got the attr access logic to work as intended.
When accessing some attr on a type, we'll try to lookup the attr in
this objects attrs. If not found, search the attr in its parent types.
And don't bind any found (and then returned) functions to this type
object instance.
This means "SomeType.some_method()" won't get "SomeType" as
first argument - which is what we want - we're accessing the function
through a type, not from an instance.

Then, if the attr was not found in this type object nor its parent types,
search through the type of this type object (its metatype/metaclass).
If we find a function, we'll bind it to this type object instance -
that's because in this case this type object really acts as an
instance of its type (metatype).

This is how Python does it - and apparently it works well.